backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[project]
name = "newsletter-backend"
version = "0.1.0"
description = "AI Newsletter backend — pipeline, API, and agent"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.0",
    "feedparser>=6.0.12",
    "firebase-admin>=6.7.0",
    "gnews>=0.4.3",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "google-cloud-firestore>=2.26.0",
    "google-cloud-storage>=3.10.1",
    "google-genai>=1.70.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdownify>=0.14.1",
    "markupsafe>=3.0.3",
    "matplotlib>=3.10.8",
    "newspaper4k>=0.9.5",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.34.0",
]
